% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{HasAttachment}
\alias{HasAttachment}
\title{Find the records which have a particular attachment.}
\usage{
HasAttachment(ob, attachment)
}
\arguments{
\item{ob}{- Data frame of imma observations.}

\item{attachment}{- Identifier of attachment to be checked ('C0', 'C1', ...'C99').}
}
\value{
for each observation, TRUE if attachment present, False otherwise
}
\description{
Assumes all records have a Core attachment
}
\details{
Used in writing IMMA - basically just to save writing null attachments
}

