% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{WhichAttachment}
\alias{WhichAttachment}
\title{Find out which attachment a parameter is in}
\usage{
WhichAttachment(parameter)
}
\arguments{
\item{parameter}{- Name of parameter to be found}
}
\value{
the number of the attachment containing that parameter.
}
\description{
IMMA data are divided into core parameters and optional attachments
this function gives the attachment number of a parameter name.
}
\details{
Throws error if parameter on tin any attachment.
}

