% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.R
\name{WriteObs}
\alias{WriteObs}
\title{Write out a data frame of IMMA records to a connection}
\usage{
WriteObs(obs, con)
}
\arguments{
\item{obs}{- Data frame of imma obs.}

\item{con}{Connection to write data to - or string, see \code{readLines()}.}
}
\description{
Operates on a data frame - like that produced by \code{ReadObs}.
}
\details{
Currently supports IMMA1 format (Rev. 17) - only main record files,
 not extension files.
}

