% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckParameter}
\alias{CheckParameter}
\title{Check the value for a parameter is inside its acceptable range(s)}
\usage{
CheckParameter(ob, parameter)
}
\arguments{
\item{ob}{Observations data frame.}

\item{parameter}{- Name of parameter to be tested}
}
\value{
for each observbation, TRUE if within range (or no range defined), FALSE if
 outside range.
}
\description{
Flags data which is physically impossible and can't be written in IMMA.
}
\details{
The IMMA format constrains the possible ranges of numeric parameters
 data outside those ranges can't be written in the format.This function
 tests all the data for a selected paramete is inside the acceptable range.
}

